Updated singer sdk and refactored tap. Added in parent child stream
