+++ pyproject.toml
@@ -1,19 +1,52 @@
 [tool.poetry]
 name = "tap-powerbi-metadata"
-version = "0.0.4"
+version = "0.1.0"
 description = "`tap-powerbi-metadata` is Singer-compliant PowerBIMetadata tap built with Singer SDK."
 authors = ["AJ Steers <aaaronsteers@gmail.com>", "John Timeus <john.timeus@slalom.com>"]
 license = "Apache v2"
 
 [tool.poetry.dependencies]
-python = ">=3.6,<3.9"
-singer-sdk = "^0.1.0"
+python = ">=3.8"
+importlib-resources = { version = "==6.1.*", python = "<3.9" }
+singer-sdk = { version="~=0.34.1" }
+fs-s3fs = { version = "~=1.1.1", optional = true }
+requests = "~=2.31.0"
 
-[tool.poetry.dev-dependencies]
-pytest = "^6.1.2"
+[tool.poetry.group.dev.dependencies]
+pytest = ">=7.4.0"
+singer-sdk = { version="~=0.34.1", extras = ["testing"] }
+
+[tool.poetry.extras]
+s3 = ["fs-s3fs"]
+
+[tool.mypy]
+python_version = "3.11"
+warn_unused_configs = true
+
+[tool.ruff]
+src = ["tap_gapi"]
+target-version = "py38"
+
+[tool.ruff.lint]
+ignore = [
+    "ANN101",  # missing-type-self
+    "ANN102",  # missing-type-cls
+    "COM812",  # missing-trailing-comma
+    "ISC001",  # single-line-implicit-string-concatenation
+]
+select = ["ALL"]
+
+[tool.ruff.lint.flake8-annotations]
+allow-star-arg-any = true
+
+[tool.ruff.lint.isort]
+known-first-party = ["tap_gapi"]
+
+[tool.ruff.lint.pydocstyle]
+convention = "google"
 
 [build-system]
-requires = ["poetry-core>=1.0.0"]
+requires = ["poetry-core==1.8.1"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
