+++ tap_powerbi_metadata/streams.py
@@ -8,9 +8,7 @@ from urllib import parse
 import requests
 import json
 
-
-from singer_sdk.streams import RESTStream
-from singer_sdk.authenticators import APIAuthenticatorBase, SimpleAuthenticator, OAuthAuthenticator, OAuthJWTAuthenticator
+from tap_powerbi_metadata.client import TapPowerBIMetadataStream, TapPowerBIUsageStream
 from singer_sdk.typing import (
     ArrayType,
     BooleanType,
@@ -24,167 +22,6 @@ from singer_sdk.typing import (
 )
 
 API_DATE_FORMAT = "'%Y-%m-%dT%H:%M:%SZ'"
-
-class OAuthActiveDirectoryAuthenticator(OAuthAuthenticator):
-    # https://pivotalbi.com/automate-your-power-bi-dataset-refresh-with-python
-
-    @property
-    def oauth_request_body(self) -> dict:
-        return {
-            'grant_type': 'password',
-            'scope': 'https://api.powerbi.com',
-            'resource': 'https://analysis.windows.net/powerbi/api',
-            'client_id': self.config["client_id"],
-            'username': self.config["username"],
-            'password': self.config["password"],
-        }
-
-class TapPowerBIMetadataStream(RESTStream):
-    """Base class for PowerBIMetadata streams."""
-
-    url_base = "https://api.powerbi.com/v1.0/myorg"
-
-    #: Variables to indicate if the endpoint requires a $top and $skip value. Defaults to False
-    top_required = False
-    skip_required = False
-
-    #: Empty dictionary that can be used if the endpoint requires additional URI parameters.
-    #: Example: uri_parameters = {'$expand':'users,reports,dashboards,datasets,dataflows,workbooks'}
-    uri_parameters = {}
-
-    #: Integer that determines the $top and $skip values for the streams that use them
-    #: 5000 is the maximum amount. Shouldn't need to be changed unless a smaller size is required.
-    top_value = 5000
-
-    def parse_response(self, response: requests.Response) -> Iterable[dict]:
-        """Parse the response and return an iterator of result rows.
-        
-        Args:
-            response: A `requests.Response` object. 
-
-        Yields:
-            One item for every item found in the response.
-        
-        """
-        resp_json = response.json()
-        for row in resp_json.get("value"):
-            yield row
-    
-    @property
-    def authenticator(self) -> APIAuthenticatorBase:
-        return OAuthActiveDirectoryAuthenticator(
-            stream=self,
-            auth_endpoint=f"https://login.microsoftonline.com/{self.config['tenant_id']}/oauth2/token",
-            oauth_scopes="https://analysis.windows.net/powerbi/api",
-        )
-
-    def get_url_params(self, partition: Optional[dict], next_page_token: Optional[Any] = None) -> Dict[str, Any]:
-        """Return a dictionary of values to be used in URL parameterization.
-        
-
-        """
-        params = self.uri_parameters
-        if self.top_required:
-            params["$top"] = self.top_value
-        if next_page_token:
-            params["$skip"] = next_page_token
-        self.logger.info(params)
-        return params
-
-    def get_next_page_token(self, response: requests.Response, previous_token: Optional[Any] = None) -> Optional[Any]:
-        """Return token for identifying next page or None if not applicable.
-        
-        Args:
-            response: A `requests.Response` object.
-            previous_token: An integer that is the $skip value that was used in the previous request.
-
-        Returns:
-            Either a new next_page_token that is the sum of the previous_token and top_value
-            Or None if a $skip value is not required, or if the 'value' of the Response object is empty
-        """
-
-        if not self.skip_required:
-            return None
-        if not previous_token:
-            next_page_token = self.top_value
-        #if there is a previous_token, but there is not content in the Response object
-        elif not response.json().get('value'):
-            return None
-        else:
-            previous_token += self.top_value
-            next_page_token = previous_token
-        return next_page_token
-
-class TapPowerBIUsageStream(RESTStream):
-    """PowerBIUsage stream class."""
-
-    url_base = "https://api.powerbi.com/v1.0/myorg"
-
-    def get_url_params(self, partition: Optional[dict], next_page_token: Optional[Any] = None) -> Dict[str, Any]:
-        """Return a dictionary of values to be used in URL parameterization.
-        
-        API only supports a single UTC day, or continuationToken-based pagination.
-        """
-        params = {}
-        if next_page_token:
-            starting_datetime = next_page_token["urlStartDate"]
-            continuationToken = next_page_token.get("continuationToken")
-        else:
-            starting_datetime = self.get_starting_timestamp(partition)
-            continuationToken = None
-        if continuationToken:
-            params["continuationToken"] = "'" + continuationToken + "'"
-        else:
-            params.update({"startDateTime": starting_datetime.strftime(API_DATE_FORMAT)})
-            ending_datetime = starting_datetime.replace(hour=0, minute=0, second=0) + timedelta(days=1) + timedelta(microseconds=-1)
-            params.update({"endDateTime": ending_datetime.strftime(API_DATE_FORMAT)})
-        self.logger.debug(params)
-        return params
-
-    @property
-    def authenticator(self) -> APIAuthenticatorBase:
-        return OAuthActiveDirectoryAuthenticator(
-            stream=self,
-            auth_endpoint=f"https://login.microsoftonline.com/{self.config['tenant_id']}/oauth2/token",
-            oauth_scopes="https://analysis.windows.net/powerbi/api",
-        )
-
-    def get_next_page_token(self, response: requests.Response, previous_token: Optional[Any] = None) -> Optional[Any]:
-        """Return token for identifying next page or None if not applicable."""
-        resp_json = response.json()
-        continuationToken = resp_json.get("continuationToken")
-        next_page_token = {}
-        if not previous_token:
-            # First time creating a pagination token so we need to record the initial start date.
-            req_url = response.request.url
-            req_params = parse.parse_qs(parse.urlparse(req_url).query)
-            self.logger.debug("Params: {}".format(req_params))
-            latest_url_start_date_param = req_params["startDateTime"][0]
-            next_page_token["urlStartDate"] = datetime.strptime(latest_url_start_date_param, API_DATE_FORMAT)
-        else: 
-            next_page_token["urlStartDate"] = previous_token.get("urlStartDate")
-        if continuationToken:
-            next_page_token["continuationToken"] = requests.utils.unquote(continuationToken)
-        else:
-            next_page_token["continuationToken"] = None
-            # Now check if we should repeat API call for next day
-            latestUrlStartDate = next_page_token["urlStartDate"]
-            nextUrlStartDate = latestUrlStartDate.replace(hour=0, minute=0, second=0) + timedelta(days=1)
-            self.logger.info("No next page token found, checking if {} is greater than now".format(nextUrlStartDate))
-            if nextUrlStartDate < datetime.utcnow():
-                self.logger.info("{} is less than now, incrementing date by 1 and continuing".format(nextUrlStartDate))
-                next_page_token["urlStartDate"] = nextUrlStartDate
-                self.logger.debug(next_page_token)
-            else:
-                self.logger.info("No continuationToken, and nextUrlStartDate after today, calling it quits")
-                return None
-        return next_page_token
-
-    def parse_response(self, response: requests.Response) -> Iterable[dict]:
-        """Parse the response and return an iterator of result rows."""
-        resp_json = response.json()
-        for row in resp_json.get("activityEventEntities"):
-            yield row
 
 class ActivityEventsStream(TapPowerBIUsageStream):
     """Returns a list of audit activity events for a tenant.
@@ -824,4 +661,34 @@ class DatasetStream(TapPowerBIMetadataStream):
         ),
         Property("webUrl", StringType),
         Property("workspaceId", StringType),
+    ).to_dict()
+
+    def get_child_context(self, record: dict, context:Optional[dict]) -> dict:
+        """ Return a context dictionary for child streams."""
+        return {
+            "datasetId": record["id"]
+        }
+class DataSourceStream(TapPowerBIMetadataStream):
+    """ Returns a list of datasets for the organization.
+    Docs: https://learn.microsoft.com/en-us/rest/api/power-bi/admin/datasets-get-datasets-as-admin
+    """
+    name = "DataSources"
+    path = "/admin/datasets/{datasetId}/datasources"
+    primary_keys = ["datasetId","datasourceId"]
+    parent_stream_type = DatasetStream
+    # top_required = True
+    # skip_required = True
+    schema = PropertiesList(
+        Property("datasourceId", StringType),
+        Property("datasetId", StringType),
+        Property("name", StringType),
+        Property("connectionString", StringType),
+        Property("gatewayId", StringType),
+        Property(
+            "connectionDetails", 
+            ObjectType(
+                Property("server", StringType),
+                Property("database", StringType),
+            )
+        ),
     ).to_dict()
